--- views/NewHome/Card.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Card = (props) =>  {
-    const showCard = props.showCard;
-}
-
-export default Card;
