+++ views/ResourcesCenter/index.ts
